+++ src/models/artifact/ArtifactSetBonus.ts
@@ -1,4 +1,4 @@
-import { JsonObject } from "config_file.js";
+import { JsonManager, JsonObject } from "config_file.js";
 import TextAssets from "../assets/TextAssets";
 import StatusProperty, { FightProp } from "../StatusProperty";
 import EnkaClient from "../../client/EnkaClient";
@@ -32,15 +32,17 @@ class ArtifactSetBonus {
 
         this._data = data;
 
-        this.id = data.affixId as number;
+        const json = new JsonManager(this._data, true, true);
+
+        this.id = json.getAsNumber("affixId");
 
         this.needCount = needCount;
 
-        this.description = new TextAssets(data.descTextMapHash as number, enka);
+        this.description = new TextAssets(json.getAsNumber("descTextMapHash"), enka);
 
-        this.addProps = (data.addProps as JsonObject[]).filter(p => Object.keys(p).includes("propType") && Object.keys(p).includes("value")).map(p => new StatusProperty(p.propType as FightProp, p.value as number, enka));
+        this.addProps = json.get("addProps").detach().filter(p => p.has("propType") && p.has("value")).map(p => new StatusProperty(p.getAsString("propType") as FightProp, p.getAsNumber("value"), enka));
 
-        this.paramList = data.paramList as number[];
+        this.paramList = json.get("paramList").detach().map(p => p.getAsNumber());
     }
 }