+++ src/models/character/talents/UpgradableSkill.ts
@@ -1,4 +1,4 @@
-import { JsonObject } from "config_file.js";
+import { JsonManager } from "config_file.js";
 import EnkaClient from "../../../client/EnkaClient";
 import SkillAttributeAssets from "../../assets/SkillAttributeAssets";
 import TextAssets from "../../assets/TextAssets";
@@ -22,15 +22,17 @@ class UpgradableSkill extends Skill {
      * @param level
      */
     getSkillAttributes(level: number): SkillAttributeAssets[] {
-        const proudSkillGroupId = this._data.proudSkillGroupId;
+        const proudSkillGroupId = new JsonManager(this._data, true, true).getAsNumber("proudSkillGroupId");
         if (!proudSkillGroupId) return [];
 
-        const leveledSkillData = this.enka.cachedAssetsManager.getGenshinCacheData("ProudSkillExcelConfigData").find(s => s.proudSkillGroupId === proudSkillGroupId && s.level === level);
+        const leveledSkillData = this.enka.cachedAssetsManager.getGenshinCacheData("ProudSkillExcelConfigData").find(p => p.getAsNumber("proudSkillGroupId") === proudSkillGroupId && p.getAsNumber("level") === level)?.detach();
         if (!leveledSkillData) return [];
 
-        if (!leveledSkillData.paramDescList) return [];
+        const paramDescList = leveledSkillData.has("paramDescList") ? leveledSkillData.get("paramDescList").detach().map(p => p.getAsNumber()) : undefined;
 
-        return (leveledSkillData.paramDescList as number[]).map(id => {
+        if (!paramDescList) return [];
+
+        return paramDescList.map(id => {
             // TODO: better filter
             try {
                 new TextAssets(id, this.enka).get("en");
@@ -38,7 +40,7 @@ class UpgradableSkill extends Skill {
                 return null;
             }
 
-            return new SkillAttributeAssets(id, this.enka, leveledSkillData.paramList as number[]);
+            return new SkillAttributeAssets(id, this.enka, leveledSkillData.has("paramList") ? leveledSkillData.get("paramList").detach().map(p => p.getAsNumber()) : []);
         }).filter(attr => attr !== null).map(attr => attr as NonNullable<typeof attr>);
     }
 
@@ -46,13 +48,13 @@ class UpgradableSkill extends Skill {
      * @param level the base level you want to upgrade to. (Do not add extra levels.)
      */
     getUpgradeCost(level: number): UpgradeCost | null {
-        const proudSkillGroupId = this._data.proudSkillGroupId;
+        const proudSkillGroupId = new JsonManager(this._data, true, true).getAsNumber("proudSkillGroupId");
         if (!proudSkillGroupId) return null;
 
-        const leveledSkillData = this.enka.cachedAssetsManager.getGenshinCacheData("ProudSkillExcelConfigData").find(s => s.proudSkillGroupId === proudSkillGroupId && s.level === level);
+        const leveledSkillData = this.enka.cachedAssetsManager.getGenshinCacheData("ProudSkillExcelConfigData").find(p => p.getAsNumber("proudSkillGroupId") === proudSkillGroupId && p.getAsNumber("level") === level)?.detach();
         if (!leveledSkillData) return null;
 
-        return new UpgradeCost((leveledSkillData.coinCost ?? 0) as number, (leveledSkillData.costItems ?? []) as JsonObject[], this.enka);
+        return new UpgradeCost(leveledSkillData.has("coinCost") ? leveledSkillData.getAsNumber("coinCost") : 0, leveledSkillData.has("costItems") ? leveledSkillData.get("costItems").detach().map(p => p.getAsJsonObject()) : [], this.enka);
     }
 }