+++ src/models/enka/CharacterBuild.ts
@@ -1,4 +1,4 @@
-import { JsonObject } from "config_file.js";
+import { JsonManager, JsonObject } from "config_file.js";
 import Character from "../character/Character";
 import EnkaClient from "../../client/EnkaClient";
 
@@ -42,17 +42,19 @@ class CharacterBuild {
 
         this.enkaUserInfo = { username: username, hash: hash };
 
-        this.id = data.id as number;
+        const json = new JsonManager(this._data, true, true);
 
-        this.name = data.name as string;
+        this.id = json.getAsNumber("id");
 
-        this.order = data.order as number;
+        this.name = json.getAsString("name");
 
-        this.isLive = data.live as boolean;
+        this.order = json.getAsNumber("order");
 
-        this.isPublic = data.public as boolean;
+        this.isLive = json.getAsBoolean("live");
 
-        this.character = new Character(data.avatar_data as JsonObject, enka);
+        this.isPublic = json.getAsBoolean("public");
+
+        this.character = new Character(json.getAsJsonObject("avatar_data"), enka);
 
         this.url = `${this.enka.options.enkaUrl}/u/${this.enkaUserInfo.username}/${this.enkaUserInfo.hash}/${this.character.characterData.id}/${this.id}`;
     }